m293+docs:等级加法整型溢出加固+三笔入册——起因=作者问2^64是不是上限;答疑=2^64(18.4Qi)是m220属性上限(double),技能书等级上限是另一个数=10亿(m149,存int,int顶21.4亿),百分比效果靠封顶封死等级再高无用、平铺类10亿级已够摸属性上限故等级不升long(大手术零收益),入DEVLOG;审出真隐患=等级加法七处裸int堆近21.4亿再加回卷负数,最狠EquipmentEnhancer.addLevels回卷被withLevel钳0=装备等级一夜清零;七处全改long运算后钳[0,intMax]:addLevels/attempt免失败直通/保护卷预判比较/安全段后到顶停推/循环level++防回卷/SkillEffectManager.learn与useAllBooks两处等级×数量(1e9×64=6.4e10乘法先溢出)/PlayerSkillManager.learnHealth;纯防御性常规数值零行为变化;零配置仍76零新API零待编译验证。docs=HANDOVER状态行推进m293+DEVLOG三条(含m290『禁疗拦得住吸血』注释纠错入册)

+++ src/main/java/com/yongye/system/PlayerSkillManager.java
@@ -32,7 +32,7 @@ public final class PlayerSkillManager {
         if (cur >= max) {
             return LearnResult.CAPPED;
         }
-        int next = Math.min(max, cur + Math.max(1, level));
+        int next = (int) Math.min(max, (long) cur + Math.max(1, level)); // m293:先 long 后钳,防溢出
         player.setAttached(ModAttachments.LEARNED_HEALTH, next);
         applyHealthModifier(player);
         com.yongye.network.YongyeNet.sendStats(player);
